working on hood tests

+++ xerolibs/xerobase/Robot.cpp
@@ -275,7 +275,7 @@ namespace xero {
                 strm << std::setfill('0') << std::setw(2) << std::hex ;                
                 if (i != 0)
                     strm << ":" ;
-                strm << macaddr_[i] ;
+                strm << (int)macaddr_[i] ;
             }
             message_logger_ << strm.str() ;
             message_logger_.endMessage() ;
